requirements.txt
streamlit
requests
google-genai
python-dotenv